\name{updateTinyMCE}
\alias{updateTinyMCE}
\title{Update tinyMCE editor}
\usage{
updateTinyMCE(session, inputId, content)
}
\arguments{
  \item{session}{the \code{session} object passed to
  function given to \code{shinyServer}}

  \item{inputId}{id associated with the tinyMCE editor}

  \item{content}{new content to place withing the editor}
}
\description{
Update tinyMCE editor object to display new content.
}

